+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### Unreleased
+
+Improvements:
+- Add autofix for `ForeignKeyTypeChecker`.
+
 ### [1.3.4] - 2022/11/12
 
 Improvements: